refactor: split bgp/peer (fsm/routes/keepalive) + bgp/manager (route)
